Reorganise some code.
