+++ build.go
@@ -3,6 +3,7 @@ package gb
 import (
 	"fmt"
 	"path/filepath"
+	"strings"
 	"time"
 )
 
@@ -18,9 +19,179 @@ func Build(pkgs ...*Package) error {
 	return Execute(build)
 }
 
-// Compile returns a Target representing all the steps required to build a go package.
-func Compile(pkg *Package, deps ...Target) PkgTarget {
-	panic("unimplemented")
+// BuildAction produces a tree of *Actions that can be executed to build
+// a *Package.
+// BuildAction walks the tree of *Packages and returns a corresponding
+// tree of *Actions representing the steps required to build *Package
+// and any of its dependencies
+func BuildAction(pkgs ...*Package) (*Action, error) {
+	targets := make(map[string]*Action) // maps package importpath ot Action name
+
+	names := func(pkgs []*Package) []string {
+		var names []string
+		for _, pkg := range pkgs {
+			names = append(names, pkg.ImportPath)
+		}
+		return names
+	}
+
+	// create top level build action to unify all packages
+	build := Action{
+		Name: fmt.Sprintf("build: %s", strings.Join(names(pkgs), ",")),
+		Task: TaskFn(func() error {
+			fmt.Println("built")
+			return nil
+		}),
+	}
+	for _, pkg := range pkgs {
+		a, err := buildAction0(targets, pkg)
+		if err != nil {
+			return nil, err
+		}
+		if a == nil {
+			// nothing to do
+			continue
+		}
+		build.Deps = append(build.Deps, a)
+	}
+	return &build, nil
+}
+
+func buildAction0(targets map[string]*Action, pkg *Package) (*Action, error) {
+
+	// if this action is already present in the map, return it
+	// rather than creating a new action.
+	if a, ok := targets[pkg.ImportPath]; ok {
+		return a, nil
+	}
+
+	// step 0. are we stale ?
+	// if this package is not stale, then by definition none of its
+	// dependencies are stale, so ignore this whole tree.
+	if !pkg.Stale {
+		return nil, nil
+	}
+
+	// step 1. walk dependencies
+	var deps []*Action
+	for _, i := range pkg.Imports() {
+		a, err := buildAction0(targets, i)
+		if err != nil {
+			return nil, err
+		}
+		if a == nil {
+			// no action required for this Package
+			continue
+		}
+		deps = append(deps, a)
+	}
+
+	// step 2. create a tree of tasks and actions for building this package.
+
+	// step 2a. are there any .s files to assemble.
+
+	var assemble []*Action
+	var ofiles []string // additional ofiles to pack
+	for _, sfile := range pkg.SFiles {
+		sfile := sfile
+		ofile := filepath.Join(pkg.Workdir(), pkg.ImportPath, stripext(sfile)+".6")
+		assemble = append(assemble, &Action{
+			Name: fmt.Sprintf("asm: %s/%s", pkg.ImportPath, sfile),
+			Task: TaskFn(func() error {
+				t0 := time.Now()
+				err := pkg.tc.Asm(pkg, pkg.Dir, ofile, filepath.Join(pkg.Dir, sfile))
+				pkg.Record("asm", time.Since(t0))
+				return err
+			}),
+		})
+		ofiles = append(ofiles, ofile)
+	}
+
+	var gofiles []string
+	gofiles = append(gofiles, pkg.GoFiles...)
+
+	// step 2b. are there any .c files that we have to run cgo on ?
+
+	if len(pkg.CgoFiles) > 0 {
+		cgoACTION, cgoOFILES, cgoGOFILES, err := cgo(pkg)
+		if err != nil {
+			return nil, err
+		}
+
+		gofiles = append(gofiles, cgoGOFILES...)
+		ofiles = append(ofiles, cgoOFILES...)
+		deps = append(deps, cgoACTION)
+	}
+
+	// step 2c. compile all the go files for this package, including pkg.CgoFiles
+
+	compile := Action{
+		Name: fmt.Sprintf("compile: %s", pkg.ImportPath),
+		Deps: deps,
+		Task: TaskFn(func() error {
+			return Compile(pkg, gofiles)
+		}),
+	}
+
+	build := &compile
+
+	// Do we need to pack ? Yes, replace build action with pack.
+	if len(ofiles) > 0 {
+		pack := Action{
+			Name: fmt.Sprintf("pack: %s", pkg.ImportPath),
+			Deps: []*Action{
+				&compile,
+			},
+			Task: TaskFn(func() error {
+				// collect .o files, ofiles always starts with the gc compiled object.
+				// TODO(dfc) objfile(pkg) should already be at the top of this set
+				ofiles = append(
+					[]string{objfile(pkg)},
+					ofiles...,
+				)
+
+				// pack
+				t0 := time.Now()
+				err := pkg.tc.Pack(pkg, ofiles...)
+				pkg.Record("pack", time.Since(t0))
+				return err
+			}),
+		}
+		pack.Deps = append(pack.Deps, assemble...)
+		build = &pack
+	}
+
+	// should this package be cached
+	// TODO(dfc) pkg.SkipInstall should become Install
+	if !pkg.SkipInstall && pkg.Scope != "test" {
+		install := Action{
+			Name: fmt.Sprintf("install: %s", pkg.ImportPath),
+			Deps: []*Action{
+				build,
+			},
+			Task: TaskFn(func() error {
+				return copyfile(pkgfile(pkg), objfile(pkg))
+			}),
+		}
+		build = &install
+	}
+
+	// if this is a main package, add a link stage
+	if pkg.isMain() {
+		link := Action{
+			Name: fmt.Sprintf("link: %s", pkg.ImportPath),
+			Deps: []*Action{build},
+			Task: TaskFn(func() error {
+				return Link(pkg)
+			}),
+		}
+		build = &link
+	}
+
+	// record the final action as the action that represents
+	// building this package.
+	targets[pkg.ImportPath] = build
+	return build, nil
 }
 
 // ObjTarget represents a compiled Go object (.5, .6, etc)
@@ -31,70 +202,39 @@ type ObjTarget interface {
 	Objfile() string
 }
 
-type gc struct {
-	target
-	pkg     *Package
-	gofiles []string
-}
-
-func (g *gc) String() string {
-	return fmt.Sprintf("compile %v", g.pkg)
-}
-
-func (g *gc) compile() error {
+func Compile(pkg *Package, gofiles []string) error {
 	t0 := time.Now()
-	if g.pkg.Scope != "test" {
+	if pkg.Scope != "test" {
 		// only log compilation message if not in test scope
-		Infof(g.pkg.ImportPath)
+		Infof(pkg.ImportPath)
 	}
-	includes := g.pkg.IncludePaths()
-	importpath := g.pkg.ImportPath
-	if g.pkg.Scope == "test" && g.pkg.ExtraIncludes != "" {
+	includes := pkg.IncludePaths()
+	importpath := pkg.ImportPath
+	if pkg.Scope == "test" && pkg.ExtraIncludes != "" {
 		// TODO(dfc) gross
-		includes = append([]string{g.pkg.ExtraIncludes}, includes...)
+		includes = append([]string{pkg.ExtraIncludes}, includes...)
 	}
-	for i := range g.gofiles {
-		if filepath.IsAbs(g.gofiles[i]) {
+	for i := range gofiles {
+		if filepath.IsAbs(gofiles[i]) {
 			// terrible hack for cgo files which come with an absolute path
 			continue
 		}
-		fullpath := filepath.Join(g.pkg.Dir, g.gofiles[i])
-		path, err := filepath.Rel(g.pkg.Projectdir(), fullpath)
+		fullpath := filepath.Join(pkg.Dir, gofiles[i])
+		path, err := filepath.Rel(pkg.Projectdir(), fullpath)
 		if err == nil {
-			g.gofiles[i] = path
+			gofiles[i] = path
 		} else {
-			g.gofiles[i] = fullpath
+			gofiles[i] = fullpath
 		}
 	}
-	err := g.pkg.tc.Gc(g.pkg, includes, importpath, g.pkg.Projectdir(), g.Objfile(), g.gofiles, g.pkg.Complete())
-	g.pkg.Record("compile", time.Since(t0))
+	err := pkg.tc.Gc(pkg, includes, importpath, pkg.Projectdir(), objfile(pkg), gofiles, pkg.Complete())
+	pkg.Record("compile", time.Since(t0))
 	return err
-}
-
-func (g *gc) Objfile() string {
-	return objfile(g.pkg)
-}
-
-func (g *gc) Pkgfile() string {
-	return g.Objfile()
 }
 
 type objpkgtarget interface {
 	ObjTarget
 	Pkgfile() string // implements PkgTarget
-}
-
-// Gc returns a Target representing the result of compiling a set of gofiles with the Context specified gc Compiler.
-func Gc(pkg *Package, gofiles []string, deps ...Target) objpkgtarget {
-	if len(gofiles) == 0 {
-		return ErrTarget{fmt.Errorf("Gc: no Gofiles provided")}
-	}
-	gc := gc{
-		pkg:     pkg,
-		gofiles: gofiles,
-	}
-	gc.target = newTarget(gc.compile, deps...)
-	return &gc
 }
 
 // PkgTarget represents a Target that produces a pkg (.a) file.
@@ -105,54 +245,22 @@ type PkgTarget interface {
 	Pkgfile() string
 }
 
-// Pack returns a Target representing the result of packing a
-// set of Context specific object files into an archive.
-func Pack(pkg *Package, deps ...ObjTarget) PkgTarget {
-	panic("removed")
-}
-
-// Asm returns a Target representing the result of assembling
-// sfile with the Context specified asssembler.
-func Asm(pkg *Package, sfile string, deps ...Target) ObjTarget {
-	panic("removed")
-}
-
-type ld struct {
-	target
-	pkg   *Package
-	afile PkgTarget
-}
-
-func (l *ld) link() error {
+func Link(pkg *Package) error {
 	t0 := time.Now()
-	target := l.pkg.Binfile()
+	target := pkg.Binfile()
 	if err := mkdir(filepath.Dir(target)); err != nil {
 		return err
 	}
 
-	includes := l.pkg.IncludePaths()
-	if l.pkg.Scope == "test" && l.pkg.ExtraIncludes != "" {
+	includes := pkg.IncludePaths()
+	if pkg.Scope == "test" && pkg.ExtraIncludes != "" {
 		// TODO(dfc) gross
-		includes = append([]string{l.pkg.ExtraIncludes}, includes...)
+		includes = append([]string{pkg.ExtraIncludes}, includes...)
 		target += ".test"
 	}
-	err := l.pkg.tc.Ld(l.pkg, includes, target, l.afile.Pkgfile())
-	l.pkg.Record("link", time.Since(t0))
+	err := pkg.tc.Ld(pkg, includes, target, objfile(pkg))
+	pkg.Record("link", time.Since(t0))
 	return err
-}
-
-// Ld returns a Target representing the result of linking a
-// Package into a command with the Context provided linker.
-func Ld(pkg *Package, afile PkgTarget) Target {
-	if !pkg.Stale {
-		return &cachedTarget{target: afile}
-	}
-	ld := ld{
-		pkg:   pkg,
-		afile: afile,
-	}
-	ld.target = newTarget(ld.link, afile)
-	return &ld
 }
 
 // objfile returns the name of the object file for this package
